fix: modify dragging logic in event handlers
add a new variable `pointerMoveActive` within the
CarouselDragEvtHandlers class. This allows to perform a check to see
whether pointermove is active when resetting first/last slide
styles (when infinite navigation is used) in the
`resetFirstSlideStylesAfterDrag` and `resetLastSlideStylesAfterDrag`
functions used in pointerup.

This check prevents the first/last slide from 'disappearing' when the
user drags twice in quick succession when going from the last slide to
the first slide or vice versa