fix: missing slide when navigating in loop
Change the dragging logic to fix an issue where when trying to drag
from the first slide to the last and then cancelling that motion and
quickly going towards the other direction would not reset the styles
for the last slide.

The same would happen when trying to drag from the
last slide to the first slide as well but in this case, it wouldn't
reset the styles for the first slide.